requirements.txt
flake8==7.3.0
pandas==2.3.2
pep8-naming==0.15.1
plotly==6.3.0
python-dotenv==1.1.1
requests==2.32.5
streamlit==1.49.1
